requirements.txt
streamlit
torch
torchvision
timm
pandas
numpy
pillow